Attempt to fix overflow

+++ RNFSManager.m
@@ -294,8 +294,8 @@ RCT_EXPORT_METHOD(readFile:(NSString *)filepath
 }
 
 RCT_EXPORT_METHOD(read:(NSString *)filepath
-                  length: (NSInteger *)length
-                  position: (NSInteger *)position
+                  length: (double *)length
+                  position: (double *)position
                   resolver:(RCTPromiseResolveBlock)resolve
                   rejecter:(RCTPromiseRejectBlock)reject)
 {
@@ -324,11 +324,11 @@ RCT_EXPORT_METHOD(read:(NSString *)filepath
     }
 
     // Seek to the position if there is one.
-    [file seekToFileOffset: (int)position];
+    [file seekToFileOffset: position];
 
     NSData *content;
-    if ((int)length > 0) {
-        content = [file readDataOfLength: (int)length];
+    if (length > 0) {
+        content = [file readDataOfLength: length];
     } else {
         content = [file readDataToEndOfFile];
     }
